+++ src/plugins/memory/memory-types.ts
@@ -15,6 +15,12 @@ export type RxStorageMemory = RxStorage<MemoryStorageInternals<any>, RxStorageMe
 };
 
 
+export type MemoryStorageInternalsByIndex<RxDocType> = {
+    index: string[];
+    docsWithIndex: DocWithIndexString<RxDocType>[];
+    getIndexableString: (docData: RxDocumentData<RxDocType>) => string
+};
+
 /**
  * The internals are shared between multiple storage instances
  * that have been created with the same [databaseName+collectionName] combination.
@@ -48,11 +54,7 @@ export type MemoryStorageInternals<RxDocType> = {
          * writes. Instead we can do a binary search over the exisiting array
          * because RxDB also knows the previous state of the document when we do a bulkWrite().
          */
-        [indexName: string]: {
-            index: string[];
-            docsWithIndex: DocWithIndexString<RxDocType>[];
-            getIndexableString: (docData: RxDocumentData<RxDocType>) => string
-        }
+        [indexName: string]: MemoryStorageInternalsByIndex<RxDocType>;
     }
 };