+++ src/rx-storage-helper.ts
@@ -154,7 +154,7 @@ export function categorizeBulkWriteRows<RxDocType>(
      * inside of the storage. Used to determine
      * which writes cause conflicts.
      */
-    docsInDb: Map<RxDocumentData<RxDocType>[StringKeys<RxDocType>], RxDocumentData<RxDocType>>,
+    docsInDb: Map<RxDocumentData<RxDocType>[StringKeys<RxDocType>] | string, RxDocumentData<RxDocType>>,
     /**
      * The write rows that are passed to
      * RxStorageInstance().bulkWrite().